.txt files to train the model and test it
First train the model on training data, a news clip on the New Deal. Then, run the model to tag the parts of speech in testdata_untagged, which is a wikipedia article on the New Deal. Finally, compare the model's results with testdata_tagged in order to measure the accuracy of the tagging model.